gui.views.timeline: make position parameter evaluation more robust
git-svn-id: svn+ssh://svn.gna.org/svn/advene/trunk@3234 3fb6f576-06d4-0310-ad11-a4b695aa9f05

+++ lib/advene/gui/views/timeline.py
@@ -155,7 +155,7 @@ class TimeLine(AdhocView):
                 # Override a potentially existing value of elements
                 elements=c.evaluateValue(v)
             elif n == 'position':
-                default_position=long(v)
+                default_position=int(float(v))
             elif n == 'zoom':
                 default_zoom=float(v)
         if ats:
